refactor: Move commands into commands lib
