+++ cadency_commands/Cargo.toml
@@ -7,11 +7,20 @@ edition = "2021"
 
 [dependencies]
 log = "0.4.14"
+reqwest = "0.11.6"
+
+[dependencies.serde]
+version = "1.0.130"
+features = ["derive"]
 
 [dependencies.serenity]
 version = "0.11.5"
 default-features = false
 features = ["client", "gateway", "rustls_backend", "model", "voice", "cache"]
+
+[dependencies.songbird]
+version = "0.3.0"
+features = ["builtin-queue"]
 
 [dependencies.cadency_core]
 path = "../cadency_core"
