+++ cadency_commands/src/pause.rs
@@ -1,6 +1,4 @@
-use crate::commands::CadencyCommand;
-use crate::error::CadencyError;
-use crate::utils;
+use cadency_core::{utils, CadencyCommand, CadencyError};
 use serenity::{
     async_trait,
     client::Context,
@@ -9,20 +7,18 @@ use serenity::{
     },
 };
 
-pub struct Stop;
+pub struct Pause;
 
 #[async_trait]
-impl CadencyCommand for Stop {
+impl CadencyCommand for Pause {
     fn name() -> &'static str {
-        "stop"
+        "pause"
     }
 
     async fn register(ctx: &Context) -> Result<Command, serenity::Error> {
         Ok(
             Command::create_global_application_command(&ctx.http, |command| {
-                command
-                    .name("stop")
-                    .description("Stop music and clean up the track list")
+                command.name("pause").description("Pause current song")
             })
             .await?,
         )
@@ -32,26 +28,38 @@ impl CadencyCommand for Stop {
         ctx: &Context,
         command: &'a mut ApplicationCommandInteraction,
     ) -> Result<(), CadencyError> {
-        debug!("Execute stop command");
+        debug!("Execute pause command");
         if let Some(guild_id) = command.guild_id {
             utils::voice::create_deferred_response(ctx, command).await?;
             let manager = utils::voice::get_songbird(ctx).await;
             if let Some(call) = manager.get(guild_id) {
                 let handler = call.lock().await;
                 if handler.queue().is_empty() {
-                    utils::voice::edit_deferred_response(ctx, command, ":x: **Nothing to stop**")
+                    utils::voice::edit_deferred_response(ctx, command, ":x: **Nothing to pause**")
                         .await?;
                 } else {
-                    handler.queue().stop();
-                    utils::voice::edit_deferred_response(
-                        ctx,
-                        command,
-                        ":white_check_mark: :wastebasket: **Successfully stopped and cleared the playlist**",
-                    )
-                    .await?;
+                    match handler.queue().pause() {
+                        Ok(_) => {
+                            utils::voice::edit_deferred_response(
+                                ctx,
+                                command,
+                                ":pause_button: **Paused**",
+                            )
+                            .await?;
+                        }
+                        Err(err) => {
+                            error!("Failed to pause: {err:?}");
+                            utils::voice::edit_deferred_response(
+                                ctx,
+                                command,
+                                ":x: **Could not pause**",
+                            )
+                            .await?;
+                        }
+                    };
                 }
             } else {
-                utils::voice::edit_deferred_response(ctx, command, ":x: **Nothing to stop**")
+                utils::voice::edit_deferred_response(ctx, command, ":x: **Nothing to pause**")
                     .await?;
             }
         } else {
