+++ cadency_commands/src/ping.rs
@@ -1,6 +1,4 @@
-use super::CadencyCommand;
-use crate::error::CadencyError;
-use crate::utils;
+use cadency_core::{utils, CadencyCommand, CadencyError};
 use serenity::{
     async_trait,
     client::Context,
