+++ cadency_commands/src/play.rs
@@ -1,7 +1,4 @@
-use crate::commands::CadencyCommand;
-use crate::error::CadencyError;
-use crate::handler::voice::InactiveHandler;
-use crate::utils;
+use cadency_core::{handler::voice::InactiveHandler, utils, CadencyCommand, CadencyError};
 use serenity::{
     async_trait,
     client::Context,
