+++ cadency_commands/src/resume.rs
@@ -1,6 +1,4 @@
-use crate::commands::CadencyCommand;
-use crate::error::CadencyError;
-use crate::utils;
+use cadency_core::{utils, CadencyCommand, CadencyError};
 use serenity::{
     async_trait,
     client::Context,
@@ -9,18 +7,20 @@ use serenity::{
     },
 };
 
-pub struct Skip;
+pub struct Resume;
 
 #[async_trait]
-impl CadencyCommand for Skip {
+impl CadencyCommand for Resume {
     fn name() -> &'static str {
-        "skip"
+        "resume"
     }
 
     async fn register(ctx: &Context) -> Result<Command, serenity::Error> {
         Ok(
             Command::create_global_application_command(&ctx.http, |command| {
-                command.name("skip").description("Skip current song")
+                command
+                    .name("resume")
+                    .description("Resume current song if paused")
             })
             .await?,
         )
@@ -37,31 +37,31 @@ impl CadencyCommand for Skip {
             if let Some(call) = manager.get(guild_id) {
                 let handler = call.lock().await;
                 if handler.queue().is_empty() {
-                    utils::voice::edit_deferred_response(ctx, command, ":x: **Nothing to skip**")
+                    utils::voice::edit_deferred_response(ctx, command, ":x: **Nothing to resume**")
                         .await?;
                 } else {
-                    match handler.queue().skip() {
+                    match handler.queue().resume() {
                         Ok(_) => {
                             utils::voice::edit_deferred_response(
                                 ctx,
                                 command,
-                                ":fast_forward: **Skipped current song**",
+                                ":play_pause: **Resumed**",
                             )
                             .await?;
                         }
                         Err(err) => {
-                            error!("Failed to skip: {err:?}");
+                            error!("Failed to resume: {err:?}");
                             utils::voice::edit_deferred_response(
                                 ctx,
                                 command,
-                                ":x: **Could not skip**",
+                                ":x: **Could not resume**",
                             )
                             .await?;
                         }
                     };
                 }
             } else {
-                utils::voice::edit_deferred_response(ctx, command, ":x: **Nothing to skip**")
+                utils::voice::edit_deferred_response(ctx, command, ":x: **Nothing to resume**")
                     .await?;
             }
         } else {
