+++ cadency_core/Cargo.toml
@@ -4,7 +4,6 @@ version = "0.2.0"
 edition = "2021"
 
 [dependencies]
-env_logger = "0.9.0"
 log = "0.4.14"
 reqwest = "0.11.6"
 thiserror = "1.0.30"
@@ -22,10 +21,6 @@ features = ["macros", "rt-multi-thread"]
 [dependencies.songbird]
 version = "0.3.0"
 features = ["builtin-queue"]
-
-[dependencies.serde]
-version = "1.0.130"
-features = ["derive"]
 
 [dev-dependencies]
 mockall = "0.11.2"
