+++ cadency_core/src/handler/command.rs
@@ -4,14 +4,11 @@ use serenity::{
     model::{application::interaction::Interaction, event::ResumedEvent, gateway::Ready},
 };
 
-use crate::commands::{
-    command_not_implemented, setup_commands, CadencyCommand, Fib, Inspire, Ping, Slap, Urban,
-};
+use crate::command::{command_not_implemented, setup_commands, CadencyCommand};
 
-use crate::commands::{Now, Pause, Play, Resume, Skip, Stop, Tracks};
 use crate::utils::set_bot_presence;
 
-pub struct Handler;
+pub(crate) struct Handler;
 
 #[async_trait]
 impl EventHandler for Handler {
@@ -31,31 +28,7 @@ impl EventHandler for Handler {
 
     async fn interaction_create(&self, ctx: Context, interaction: Interaction) {
         if let Interaction::ApplicationCommand(mut command) = interaction {
-            let cmd_execution = match command.data.name.as_str() {
-                "ping" => Ping::execute(&ctx, &mut command).await,
-                "inspire" => Inspire::execute(&ctx, &mut command).await,
-                "fib" => Fib::execute(&ctx, &mut command).await,
-                "urban" => Urban::execute(&ctx, &mut command).await,
-                "slap" => Slap::execute(&ctx, &mut command).await,
-
-                "play" => Play::execute(&ctx, &mut command).await,
-
-                "now" => Now::execute(&ctx, &mut command).await,
-
-                "skip" => Skip::execute(&ctx, &mut command).await,
-
-                "pause" => Pause::execute(&ctx, &mut command).await,
-
-                "resume" => Resume::execute(&ctx, &mut command).await,
-
-                "stop" => Stop::execute(&ctx, &mut command).await,
-
-                "tracks" => Tracks::execute(&ctx, &mut command).await,
-                _ => command_not_implemented(&ctx, command).await,
-            };
-            if let Err(execution_err) = cmd_execution {
-                error!("❌ Command execution failed: {:?}", execution_err);
-            }
+            unimplemented!("Execute command by name");
         };
     }
 }
