+++ cadency_core/src/handler/mod.rs
@@ -1,3 +1,3 @@
-pub mod command;
+pub(crate) mod command;
 
 pub mod voice;
